Dialogue System Implemented
To get an entity to speak they must have the Actor script on them. You specify the Name, the Position where the speak indicator will appear and the dialogue tree.
Dialogue Tree can be created by right clicking somewhere in the Asset folder and going to: Create -> Dialogue -> Dialogue Asset